fix(web): fully paginate agent vaults
