+++ apps/web/src/components/dashboard/agent-vaults-tab.tsx
@@ -4,7 +4,7 @@ import { ApiErrorPanel } from "@/components/api-error-panel";
 import { useAgentProjectBindings } from "@/components/dashboard/agent-projects-tab";
 import { Skeleton } from "@/components/ui/skeleton";
 import { effectiveAgentProjectIds } from "@/components/vault/vault-scope";
-import { VaultsSurface } from "@/pages/dashboard/vault/page";
+import { VaultsSurface } from "@/components/vault/vaults-surface";
 
 export function AgentVaultsTab({ agentId }: { agentId: string }) {
 	const bindings = useAgentProjectBindings(agentId);
