+++ apps/web/src/components/vault/vault-scope.test.ts
@@ -1,6 +1,10 @@
 import { describe, expect, test } from "bun:test";
 import type { components } from "@/lib/api-schemas";
-import { effectiveAgentProjectIds, vaultsForProjectIds } from "./vault-scope";
+import {
+	effectiveAgentProjectIds,
+	fetchAgentProjectVaults,
+	vaultsForProjectIds,
+} from "./vault-scope";
 
 type Binding = components["schemas"]["AgentProjectBindingResponse"];
 type Vault = components["schemas"]["VaultResponse"];
@@ -59,5 +63,65 @@ describe("agent Vault scope", () => {
 
 		expect(visible.map((item) => item.id)).toEqual(["primary-vault", "context-vault"]);
 		expect(vaultsForProjectIds([vault("private", ["project_other"])], [])).toEqual([]);
+	});
+
+	test("walks every server-filtered page for each effective Project", async () => {
+		const calls: Array<{ projectId: string; page: number; pageSize: number }> = [];
+		const result = await fetchAgentProjectVaults(
+			["project_primary"],
+			async (projectId, page, pageSize) => {
+				calls.push({ projectId, page, pageSize });
+				return {
+					items:
+						page === 1
+							? [vault("first", [projectId]), vault("second", [projectId])]
+							: [vault("target-after-first-page", [projectId])],
+					total: 3,
+					page,
+					page_size: pageSize,
+				};
+			},
+			{ pageSize: 2 },
+		);
+
+		expect(calls).toEqual([
+			{ projectId: "project_primary", page: 1, pageSize: 2 },
+			{ projectId: "project_primary", page: 2, pageSize: 2 },
+		]);
+		expect(result.map((item) => item.id)).toEqual(["first", "second", "target-after-first-page"]);
+	});
+
+	test("deduplicates a shared Vault across Project-filtered inventories", async () => {
+		const result = await fetchAgentProjectVaults(
+			["project_primary", "project_context"],
+			async (projectId, page, pageSize) => ({
+				items:
+					projectId === "project_primary"
+						? [vault("shared", [projectId]), vault("primary-only", [projectId])]
+						: [vault("shared", [projectId])],
+				total: projectId === "project_primary" ? 2 : 1,
+				page,
+				page_size: pageSize,
+			}),
+			{ pageSize: 2 },
+		);
+
+		expect(result.map((item) => item.id)).toEqual(["shared", "primary-only"]);
+		expect(result[0]?.project_ids).toEqual(["project_primary", "project_context"]);
+	});
+
+	test("throws instead of returning a truncated Agent Vault inventory", async () => {
+		await expect(
+			fetchAgentProjectVaults(
+				["project_primary"],
+				async (projectId, page, pageSize) => ({
+					items: [vault(`page-${page}`, [projectId])],
+					total: 2,
+					page,
+					page_size: pageSize,
+				}),
+				{ pageSize: 1, maxPages: 1 },
+			),
+		).rejects.toThrow("Too many agent Vault pages");
 	});
 });
